+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Support a new error code `outdatedFundIdInEncumbrance`. Refs UINV-486.
 * Prevent approval or payment of invoice with linked order status of "pending". Refs UINV-465.
 * "Release encumbrance" checkbox becomes checked after adding new blank invoice line. Refs UINV-484.
+* Allow user to select Fund and Expense class from Fiscal year specified in invoice. Refs UINV-479.
 
 ## [4.0.1](https://github.com/folio-org/ui-invoice/tree/v4.0.1) (2023-03-10)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v4.0.0...v4.0.1)